renames samples/ to applications/ and Refactor childbenefitsclaim to  applications/
