+++ src/applications/AppSelector/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Switch, Route } from 'react-router-dom';
-import ChildBenefitsClaim from '../../journeys/ChildBenefitsClaim'
+import ChildBenefitsClaim from '../ChildBenefitsClaim'
 
 // NOTE: You should update this to be the same value that's in
 //  the src/index.html <base href="value"> to allow the React Router
