+++ src/index.tsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { render } from "react-dom";
 import { BrowserRouter } from 'react-router-dom';
-import TopLevelApp from '../src/samples/TopLevelApp';
+import TopLevelApp from './applications/TopLevelApp';
 
 const outletElement = document.getElementById("outlet");
